requirements.txt
anyio==4.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
dataclasses==0.6
Django==4.1.13
django-bootstrap-daterangepicker==1.1.0
django-clearcache==1.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
Faker==22.6.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
motor==3.3.2
mysql-connector-python==8.3.0
oscrypto==1.3.0
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
pycparser==2.21
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.15
